Scheme - scheme creation with a root path
The root path was being passed in and we were finding the shared and user
specific schemes. From within the scheme we were having to backtrack a
large number of directories to find the root again (plus one more out)
to find the project.

Instead we are just passing in the root when we create the scheme.

Cleaned up some of the methods to generate the schemes and scheme paths